requirements.txt
pytest
deepdiff
pillow
openai
transformers
torch
streamlit
bitsandbytes
accelerate
flash-attn
fastapi[standard]
